requirements.txt
dash
plotly
pandas
firebase-admin
gunicorn
python-dotenv
numpy